Introduce data directory for better docker support.
Currently dynamic data like transformation cache and the index
must be kept inside the docker container they are situated
directly under the components directory. The data lies
in directories that get renamed by the harvesting process.
Therefore they cannot be mapped by a docker volume (cannot
rename a docker volume).

This solutions introduces a data directory parallel to the
instance directory. The migration is extended to reflect the
changes in directory structure.