Map GML3.2 elements back to GML (#2505)
ATTN: THE NAME OF THE XSLT FILE HAS CHANGED!

This is a backport for the changes in 5.8.1 to the GML namespace. All occurrences
of namespace "http://www.opengis.net/gml/3.2" are set to "http://www.opengis.net/gml".

ATTN: Existing elements of namespace "http://www.opengis.net/gml" in the source are
ignored!

see also #1876